// src/enums/errors.ts
export enum ErrorsEnum {
  USER_NOT_FOUND = 'Usuario não encontrado',
  TOKEN_NOT_FOUND = 'Token não encontrado',
  TOKEN_INVALID = 'Token inválido',
  USER_ALREADY_EXISTS = 'Usuário já existe',
  INVALID_PASS_OR_EMAIL = 'Email ou senha inválidos',
  ERROR_SEND_EMAIL = 'Erro ao enviar e-mail',
  INVALID_CODE = 'Código inválido',
}